Put module in top folder
