--- python/mage/json_util/__init__.py
@@ -1,3 +1,0 @@
-"""json_util package initialization."""
-# These functions are used by external callers and tests
-from .core import to_json, from_json_list, load_from_path, load_from_url  # noqa: F401
